requirements.txt
openai-whisper
setuptools-rust
pydub==0.25.1
